com.ibm.ws.microprofile.metrics.3.x.monitor_fat bundle modified to address tests for mpMetrics-3.0.
Tests EDF1 and EF12 removed as they are no longer relevant with the mpMetrics-3.0 feature.

Test EDF1 tests the mpMetric feature without the monitor feature and expects no vendor metrics to be available. however, has hard dependency on monitor-1.0 and  thus this test is not useful at all.

Test EDF12 tests the removal of monitor-1.0 feature and expected the removal of vendor metrics.  mpMetrics-3.0, however, has hard dependency on monitor-1.0 and  thus this test is not useful at all.

The relevant files that were designated for tests EDF1 and EDF12 were removed.

Monitor-Metrics bundle had to modify MonitorMetrics.nlsprops for appropriate message code.